requirements.txt
# PDF Processing
PyMuPDF==1.23.8

# LLM Integration
openai==1.3.0

# Knowledge Graph
neo4j==5.14.1

# Data Processing
pandas==2.1.3
numpy==1.24.3

# Web API
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# Additional utilities
python-dotenv==1.0.0
tqdm==4.66.1
scikit-learn
networkx
scipy
joblib
threadpoolctl
